% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_accepted_taxa}
\alias{mnnpc_accepted_taxa}
\title{Accepted taxon names}
\format{
A data frame with 2918 rows and 1 columns, the definitions of which are:
\describe{
\item{taxon_name}{Taxon names.}
}
}
\usage{
mnnpc_accepted_taxa
}
\description{
A data frame containing the accepted taxon names following MNTaxa: The State of Minnesota Vascular Plant Checklist
as present in the \code{MNNPC::mnnpc_taxonomic_backbone} object.
}
\details{
\code{mnnpc_accepted_taxa}
}
\keyword{datasets}
