% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_bb_conv}
\alias{mnnpc_bb_conv}
\title{mnnpc_bb_conv}
\format{
A data frame with 7 rows and 2 columns, the definitions of which are:
\describe{
\item{scov}{Braun-Blanquet cover/abundance scale values.}
\item{scov_mid}{Percent cover values.}
}
}
\usage{
mnnpc_bb_conv
}
\description{
A dataframe used to convert cover values from the Braun-Blanquet cover/abundance scale to percent cover based on approximate mid-points (Tüxen and Ellenberg 1937, van der Maarel 2007).
}
\details{
\code{mnnpc_bb_conv}
}
\keyword{datasets}
