% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_development_data}
\alias{mnnpc_development_data}
\title{Floristic Table Data}
\format{
A data frame with 203232 rows and 20 columns, the definitions of which are:
\describe{
\item{Quadrat}{The ID/releve number of the plot (true numbers have been masked for data security).}
\item{Species}{The formatted taxon name (analysis group and strata, if applicable), see \code{MNNPC::mnnpc_taxa_lookup}.}
\item{Cover}{The percent cover of the taxon in the plot within the strata.}
\item{analysis_group}{The analysis group, see \code{MNNPC::mnnpc_taxa_lookup}.}
\item{strata_lower}{The minimum height class of the strata.}
\item{strata_upper}{The maximum height class of the strata.}
\item{code_strata}{Equivalent to Species, but with spaces replaced by underscores.}
\item{used_in_fieldguide}{Indicator for whether the plot was used to inform MN NPC field guides published in 2003 and 2005.}
\item{npc_code}{The most specific MN NPC unit code that the plot was identified to.}
\item{npc_system_id}{Code for the system level of the MN NPC classification system.}
\item{npc_system}{Name of the system level of the MN NPC classification system.}
\item{npc_sys_flor}{Code for the system level and floristic region of the MN NPC classification system.}
\item{npc_class}{Code for the class level of the MN NPC classification system.}
\item{npc_class_name}{Name of the class level of the MN NPC classification system.}
\item{npc_type}{Code for the type level of the MN NPC classification system, if available.}
\item{npc_type_name}{Name of the type level of the MN NPC classification system, if available.}
\item{npc_subtype}{Code for the subtype level of the MN NPC classification system, if available.}
\item{npc_subtype_name}{Name of the subtype level of the MN NPC classification system, if available.
\item{ecs_secname}{Name of the ECS section where the plot occurs.}
\item{ecs_section}{Code of the ECS section where the plot occurs.}
}
}
\usage{
mnnpc_development_data
}
\description{
The data used to build floristic tables for each MN NPC unit.
}
\details{
\code{mnnpc_development_data}
}
\keyword{datasets}
