% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_diversity_metrics}
\alias{mnnpc_diversity_metrics}
\title{Diversity statistics}
\format{
A data frame with 312525 rows and 11 columns, the definitions of which are:
\describe{
\item{ecs_section}{The abbreviation of the ECS Section in which the plots occurred, one of: Lake Agassiz, Aspen Parklands (LAP), Minnesota & NE Iowa Morainal (MIM), N. Minnesota & Ontario Peatlands(MOP), N. Minnesota Drift & Lake Plains (MDL), North Central Glaciated Plains (CGP), Northern Superior Uplands (NSU), Paleozoic Plateau (PPL), Red River Valley (RRV), Southern Superior Upland (SSU), Western Superior Uplands (WSU), or Statewide (statewide).}
\item{rank}{The MNNPC unit rank associated with the npc_class, npc_type, or npc_subtype.}
\item{npc_class}{The names of the MNNPC unit classes, the highest rank.}
\item{npc_type}{The names of the MNNPC unit types.}
\item{npc_subtype}{The names of the MNNPC unit subtypes.}
\item{Quadrat}{The name of the quadrat/releve.}
\item{measure}{The name of the diversity measure, one of: normalised alpha, normalised beta, and gamma.}
\item{metric}{The name of the diversity metric, one of: naive, taxonomic, and phylogenetic.}
\item{partition_level}{The diversity partition level, which for this object is equal to metacommunity for rows containing data for MNNPC class, type, or subtypes; for quadrats/releves the partition level is equal to subcommunity.}
\item{q}{The Hill number, for this dataset three Hill numbers were used: 0, 1, and 2.}
\item{diversity}{The diversity value.}
}
}
\usage{
mnnpc_diversity_metrics
}
\description{
Diversity statistics for the MNNPC calculated using the R package \code{rdiversity}.
}
\details{
\code{mnnpc_diversity_metrics}
}
\references{
Reeve, R., Leinster, T., Cobbold, C.A., Thompson, J., Brummitt, N., Mitchell, S.N., Matthews, L., 2016. How to partition diversity. https://doi.org/10.48550/arXiv.1404.6520
}
\keyword{datasets}
