% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_dom_conv}
\alias{mnnpc_dom_conv}
\title{mnnpc_dom_conv}
\format{
A data frame with 11 rows and 2 columns, the definitions of which are:
\describe{
\item{scov}{Domin cover/abundance scale values.}
\item{scov_mid}{Percent cover values.}
}
}
\usage{
mnnpc_dom_conv
}
\description{
A dataframe used to convert cover values from the Domin cover/abundance scale to percent cover based on approximate mid-points.
}
\details{
\code{mnnpc_dom_conv}
}
\keyword{datasets}
