% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_ecs_sections}
\alias{mnnpc_ecs_sections}
\title{MNNPC ECS Sections}
\format{
A named vector of strings of length 11.
}
\usage{
mnnpc_ecs_sections
}
\description{
A vector of named ECS section codes with full names.
}
\details{
\code{mnnpc_ecs_sections}
}
\keyword{datasets}
