% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_fidelity_metrics}
\alias{mnnpc_fidelity_metrics}
\title{Fidelity measures}
\format{
A list of 9 data frames with 733026 rows and  columns, the definitions of which are:
\describe{
\item{ecs_section}{The abbreviation of the ECS Section in which the plots occurred, one of: Lake Agassiz, Aspen Parklands (LAP), Minnesota & NE Iowa Morainal (MIM), N. Minnesota & Ontario Peatlands(MOP), N. Minnesota Drift & Lake Plains (MDL), North Central Glaciated Plains (CGP), Northern Superior Uplands (NSU), Paleozoic Plateau (PPL), Red River Valley (RRV), Southern Superior Upland (SSU), Western Superior Uplands (WSU), or Statewide (statewide).}
\item{system}{The MNNPC unit system code.}
\item{npc_class}{The MNNPC unit class code.}
\item{species}{The name of the species/taxon.}
\item{u}{The fidelity measure u, more specifically Bruelheide's u.}
\item{phi}{The fidelity measure phi.}
\item{chisq}{The fidelity measure chisq.}
\item{g}{The fidelity measure g.}
\item{indval}{The fidelity measure indval.}
}
}
\usage{
mnnpc_fidelity_metrics
}
\description{
Statistical measures of fidelity for taxa present in the MNNPC classification,
calculated following Chytrý et al (2002).
}
\details{
\code{mnnpc_fidelity_metrics}
}
\references{
Chytrý, M., Tichý, L., Holt, J., Botta-Dukát, Z., 2002. Determination of diagnostic species with statistical fidelity measures. Journal of Vegetation Science 13, 79–90. https://doi.org/10.1111/j.1654-1103.2002.tb02025.x
}
\keyword{datasets}
