% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_ht_conv}
\alias{mnnpc_ht_conv}
\title{mnnpc_ht_conv}
\format{
A data frame with 8 rows and 3 columns, the definitions of which are:
\describe{
\item{ht}{The height category to assign for strata between ht_min_m and ht_max_m.}
\item{ht_min_m}{...}
\item{ht_max_m}{...}
}
}
\usage{
mnnpc_ht_conv
}
\description{
...
}
\details{
\code{mnnpc_ht_conv}
}
\keyword{datasets}
