% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_phlyo_tree}
\alias{mnnpc_phlyo_tree}
\title{A phylogenetic tree for the MNNPC taxa}
\format{
A string containing the phylogentic tree in the Newick format.
}
\usage{
mnnpc_phlyo_tree
}
\description{
A phylogenetic tree in the Newick format, formed from taxa present in the search_name column in \code{MNNPC::mnnpc_phylo_taxa_lookup}, where available.
Branch lengths were calculated using the 'Grafen' method implemented in the \code{ape::compute.brlen} function.
}
\details{
\code{mnnpc_phlyo_tree}
}
\keyword{datasets}
