% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_pquads}
\alias{mnnpc_pquads}
\title{Pseudo-quadrats}
\format{
A data frame with 254308 rows and 3 columns, the definitions of which are:
\describe{
\item{npc_code}{The code for the MNNPC unit.}
\item{psq_id}{The pseudo-quadrat ID.}
\item{taxon_name}{The taxon name, see \code{MNNPC::mnnpc_accepted_taxa} and \code{MNNPC::mnnpc_taxonomic_backbone}.}
}
}
\usage{
mnnpc_pquads
}
\description{
Pseudo quadrats for each MNNPC unit.
}
\details{
\code{mnnpc_pquads}
}
\keyword{datasets}
