% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_releves}
\alias{mnnpc_releves}
\title{MN NPC Releves}
\format{
A data frame with 406374 rows and 3 columns, the definitions of which are:
\describe{
\item{npc_code}{The code for the MN NPC unit.}
\item{releve}{An anonymised ID for each releve.}
\item{taxon_name}{The taxon name, see \code{MNNPC::mnnpc_accepted_taxa} and \code{MNNPC::mnnpc_taxonomic_backbone}.}
}
}
\usage{
mnnpc_releves
}
\description{
The releves used to define each MNNPC unit, by ECS section, produced using the
\code{MNNPC::mnnpc_development_data} object.
}
\details{
\code{mnnpc_releves}
}
\keyword{datasets}
