% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_taxa_conv}
\alias{mnnpc_taxa_conv}
\title{Taxon name conversion}
\format{
A data frame with 8483 rows and 2 columns, the definitions of which are:
\describe{
\item{taxon}{The taxon name, see \code{MNNPC::taxa_lookup}}
\item{analysis_group}{The taxon or group of taxa used for analysis in the MNNPC and RMAVIS.}
}
}
\usage{
mnnpc_taxa_conv
}
\description{
A data frame containing all taxon names present in \code{MNNPC::taxa_lookup} and
the taxa and taxon groups that should be used for MNNPC analysis. For example,
Oxalis stricta, Oxalis dillenii, Oxalis montana, and Oxalis acetosella are all
lumped into the 'Oxalis ecological group' analysis group.
The \code{MNNPC::mnnpc_pquads}, \code{MNNPC::mnnpc_floristic_tables}, and \code{MNNPC::example_data}
objects only include taxa and taxon froups from the 'analysis_group' column
in this object and \code{RMAVIS} will only match based on these groups.
}
\details{
\code{mnnpc_taxa_conv}
}
\keyword{datasets}
