% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_taxa_lookup}
\alias{mnnpc_taxa_lookup}
\title{A lookup between the accepted MNTaxa taxa and their synonyms}
\format{
A data frame with 7781 rows and 2 columns, the definitions of which are:
\describe{
\item{taxon_name}{The taxon concept name, includes both accepted/recommended taxa and their synonyms.}
\item{recommended_taxon_name}{The accepted/recommended taxon name, see \code{MNNPC::mnnpc_accepted_taxa} and \code{MNNPC::mnnpc_taxonomic_backbone}.}
}
}
\usage{
mnnpc_taxa_lookup
}
\description{
A lookup between the accepted MNTaxa taxa and their synonyms, with additional information.
}
\details{
\code{mnnpc_taxa_lookup}
}
\keyword{datasets}
