% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_taxonomic_backbone}
\alias{mnnpc_taxonomic_backbone}
\title{The taxonomic backbone for the MNNPC}
\format{
A data frame with 2918 rows and 19 columns, the definitions of which are:
\describe{
\item{taxon_name}{The taxon concept name, includes accepted/recommended taxa only.}
\item{taxa_included}{The taxon names of the taxa included in taxon_name.}
\item{taxa_ids}{The ids of the taxa in taxa_included.}
\item{rank}{The taxonomic rank of the taxon recorded in taxon_name.}
\item{hybrid}{The hybrid status of the taxon_name: 0 = non-hybrid, 1 = hyrbid. No hybrids at present.}
\item{full_name}{The full name of the taxon concept, which is equivalent to a concatentation of the taxon_name, qualifier, and authority.}
\item{publication}{The parent publication containing the taxon_name description.}
\item{physiognomy}{The lifeform associated with the taxa_included.}
\item{common_name}{The common name associated with the taxa_included.}
\item{c_value}{The coefficients of conservatism associated with the taxa_included.}
\item{origin}{The nativeness status of the taxon_name.}
\item{species}{The name of the species taxon associated with the taxon_name.}
\item{genus}{The name of the parent genus taxon associated with the taxon_name.}
\item{family}{The name of the parent family taxon associated with the taxon_name.}
\item{order}{The name of the parent order taxon associated with the taxon_name.}
\item{class}{The name of the parent class taxon associated with the taxon_name.}
\item{phylum}The name of the parent phylum taxon associated with the taxon_name.}
\item{kingdom}{The name of the parent kingdom taxon associated with the taxon_name.}
\item{lineage_source}{The source of the parent taxa }
}
}
\usage{
mnnpc_taxonomic_backbone
}
\description{
The taxonomic backbone for the MNNPC, containin the recommended/accepted taxon names and associated taxonomic information.
}
\details{
\code{mnnpc_taxonomic_backbone}
}
\keyword{datasets}
