% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_vc_systems}
\alias{mnnpc_vc_systems}
\title{MNNPC system codes}
\format{
A vector of strings of length 12.
}
\usage{
mnnpc_vc_systems
}
\description{
A vector of codes for the MNNPC community systems which are prefixes for all child community classes, types, and sub-types.
}
\details{
\code{mnnpc_vc_systems}
}
\keyword{datasets}
