% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_vc_systems_flreg}
\alias{mnnpc_vc_systems_flreg}
\title{MNNPC systems with floristic regions}
\format{
A vector of strings of length 30.
}
\usage{
mnnpc_vc_systems_flreg
}
\description{
A vector of codes consisting of the MNNPC community systems and floristic regions.
}
\details{
\code{mnnpc_vc_systems_flreg}
}
\keyword{datasets}
