% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_vc_systems_flreg_named}
\alias{mnnpc_vc_systems_flreg_named}
\title{MNNPC systems with floristic regions with full names}
\format{
A named vector of strings of length 30.
}
\usage{
mnnpc_vc_systems_flreg_named
}
\description{
A named vector of MNNPC systems and floristic regions with full names.
}
\details{
\code{mnnpc_vc_systems_flreg_named}
}
\keyword{datasets}
