% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_vc_systems_flreg_nested}
\alias{mnnpc_vc_systems_flreg_nested}
\title{MNNPC systems by region}
\format{
A list of length 42
}
\usage{
mnnpc_vc_systems_flreg_nested
}
\description{
A list of MNNPC systems and system-regions with list entried equal to all child system-regions.
}
\details{
\code{mnnpc_vc_systems_flreg_nested}
}
\keyword{datasets}
