% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_vc_systems_named}
\alias{mnnpc_vc_systems_named}
\title{MNNPC system codes with full names}
\format{
A named vector of strings of length 12.
}
\usage{
mnnpc_vc_systems_named
}
\description{
A vector of named codes for the MNNPC community system which are prefixes for all child community classes, types, and sub-types.
}
\details{
\code{mnnpc_vc_systems_named}
}
\keyword{datasets}
