% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_vc_types}
\alias{mnnpc_vc_types}
\title{MNNPC system code types/prefixes}
\format{
A data frame with 12 which contains ...
}
\usage{
mnnpc_vc_types
}
\description{
A vector of codes for the MNNPC community system which are prefixes for all child community classes, types, and sub-types.
}
\details{
\code{mnnpc_vc_types}
}
\keyword{datasets}
