% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dnr_releves.R
\name{process_dnr_releves}
\alias{process_dnr_releves}
\title{Process releve data into the format required by RMAVIS}
\usage{
process_dnr_releves(
  releve_data,
  process_malformed_data = TRUE,
  strip_suffixes = TRUE,
  match_to_accepted = TRUE,
  aggregate_into_analysis_groups = TRUE,
  cover_scale = "percentage"
)
}
\arguments{
\item{releve_data}{See \code{MNNPC::mnnpc_example_releve}.}

\item{process_malformed_data}{Specify whether to process data with missing values, a boolean (TRUE/FALSE).}

\item{strip_suffixes}{Specify whether to strip suffixes (e.g. s.l., s.s., and s.a.) from taxon names, a boolean (TRUE/FALSE).}

\item{match_to_accepted}{Specify whether to convert raw taxon names to accepted/recommended taxon names, a boolean (TRUE/FALSE).}

\item{aggregate_into_analysis_groups}{Specify whether to aggregate taxa into analysis groups, a boolean (TRUE/FALSE).}

\item{cover_scale}{Specify the scale of species cover values, one of: "percentage", "proportional", "domin", "braunBlanquet", or "none".}
}
\value{
A data frame containing the releve data with five columns: "Year", "Group", "Quadrat", "Species", "Cover"
}
\description{
Process releve data into the format required by RMAVIS, specifically a data frame
in long format containing "Year", "Group", "Quadrat", "Species", and "Cover" columns.
}
\examples{
MNNPC::process_dnr_releves(releve_data = MNNPC::mnnpc_example_releve)
}
