% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_releve}
\alias{example_releve}
\title{example_releve}
\format{
A data frame with 86 rows and 8 columns, the definitions of which are:
\describe{
\item{year}{...}
\item{group}{...}
\item{relnumb}{...}
\item{physcode}{...}
\item{minht}{...}
\item{maxht}{...}
\item{taxon}{...}
\item{scov}{...}
}
}
\usage{
example_releve
}
\description{
...
}
\details{
\code{example_releve}
}
\keyword{datasets}
