% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_example_data}
\alias{mnnpc_example_data}
\title{Example vegetation plot data}
\format{
A list of 1 data frames with  rows and  columns, the definitions of which are:
\describe{
\item{Site}{The name of the site.}
\item{Year}{The year the quadrat was sampled in.}
\item{Group}{The group containing the quadrat.}
\item{Quadrat}{The quadrat name.}
\item{Species}{The taxon name, see \code{MNNPC::mnnpc_accepted_taxa} and \code{MNNPC::mnnpc_taxonomic_backbone}.}
\item{Cover}{The cover of the taxon in the quadrat.}
}
}
\usage{
mnnpc_example_data
}
\description{
A selection of example vegetation plot datasets, currently included are:
}
\details{
\itemize{
\item St. Croix State Forest
}

\code{mnnpc_example_data}
}
\keyword{datasets}
