% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_scov_conv}
\alias{mnnpc_scov_conv}
\title{mnnpc_scov_conv}
\format{
A data frame with 7 rows and 2 columns, the definitions of which are:
\describe{
\item{scov}{...}
\item{scov_mid}{...}
}
}
\usage{
mnnpc_scov_conv
}
\description{
...
}
\details{
\code{mnnpc_scov_conv}
}
\keyword{datasets}
