% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnnpc_strata}
\alias{mnnpc_strata}
\title{mnnpc_strata}
\format{
A data frame with 14 rows and 3 columns, the definitions of which are:
\describe{
\item{physcode}{...}
\item{strata_lower}{...}
\item{strata_upper}{...}
}
}
\usage{
mnnpc_strata
}
\description{
...
}
\details{
\code{mnnpc_strata}
}
\keyword{datasets}
