% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dnr_releves.R
\name{process_dnr_releves}
\alias{process_dnr_releves}
\title{Process releve data into the format required by RMAVIS}
\usage{
process_dnr_releves(releve_data)
}
\arguments{
\item{releve_data}{See \code{MNNPC::example_releve}.}
}
\value{
A data frame containing the releve data with five columns: "Year", "Group", "Quadrat", "Species", "Cover"
}
\description{
Process releve data into the format required by RMAVIS, specifically a data frame
in long format containing "Year", "Group", "Quadrat", "Species", and "Cover" columns.
}
\examples{
MNNPC::process_dnr_releves(releve_data = MNNPC::example_releve)
}
