% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dnr_releves.R
\name{process_dnr_releves}
\alias{process_dnr_releves}
\title{Process releve data into the format required by RMAVIS}
\usage{
process_dnr_releves(
  releve_data,
  strip_suffixes = TRUE,
  match_to_accepted = TRUE,
  aggregate_into_assigned = FALSE,
  aggregate_into_analysis_groups = TRUE,
  include_strata = TRUE,
  cover_scale = "percentage"
)
}
\arguments{
\item{releve_data}{See \code{MNNPC::mnnpc_example_releve}.}

\item{strip_suffixes}{Specify whether to strip suffixes (e.g. s.l., s.s., and s.a.) from taxon names, a boolean (TRUE/FALSE).}

\item{match_to_accepted}{Specify whether to convert raw taxon names to accepted (recommended_taxon_name in \code{MNNPC::mnnpc_taxa_lookup}), a boolean (TRUE/FALSE).}

\item{aggregate_into_assigned}{Specify whether to aggregate taxa into accepted groups (recommended_assignment in \code{MNNPC::mnnpc_taxa_lookup}), which are equivalent taxa due to taxonomic changes, a boolean (TRUE/FALSE). Ignored when aggregate_into_analysis_groups is TRUE.}

\item{aggregate_into_analysis_groups}{Specify whether to aggregate taxa into analysis groups (analysis_group in \code{MNNPC::mnnpc_taxa_lookup}), a boolean (TRUE/FALSE).}

\item{include_strata}{Specify whether to stratify tree taxa into understory, subcanopy, and canopy based on height values, a boolean (TRUE/FALSE). Can be combined with match_to_accepted, aggregate_into_assigned, or aggregate_into_analysis_groups output.}

\item{cover_scale}{Specify the scale of species cover values, one of: "percentage", "proportional", "domin", "braunBlanquet", or "none".}
}
\value{
A data frame containing the releve data with five columns: "Year", "Group", "Quadrat", "Species", "Cover"
}
\description{
Process releve data into the format required by RMAVIS, specifically a data frame
in long format containing "Year", "Group", "Quadrat", "Species", and "Cover" columns.
}
\examples{
MNNPC::process_dnr_releves(releve_data = MNNPC::mnnpc_example_releve, cover_scale = "braunBlanquet")
}
